updated aws_ec2_create_ami_from_instance.sh

+++ aws/aws_ec2_create_ami_from_instance.sh
@@ -63,6 +63,9 @@ instance_name="$1"
 ami_name="$2"
 
 instance_id="$(VERBOSE=1 "$srcdir/aws_ec2_instance_name_to_id.sh" "$instance_name")"
+if ! is_instance_id "$instance_id"; then
+    die "Invalid Instance ID returned, failed regex validation: $instance_id"
+fi
 echo >&2
 
 no_reboot="--no-reboot"
